revert state column badge changes
