+++ src/Aspire.Dashboard/Aspire.Dashboard.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <!-- This list of runtimes needs to match the set of runtimes specified in eng/workloads/workloads.csproj -->
@@ -173,6 +173,14 @@
     <Compile Include="..\Aspire.Hosting\Utils\IEnvironmentVariables.cs" Link="Utils\IEnvironmentVariables.cs" />
     <Compile Include="..\Aspire.Hosting\Utils\KnownResourceNames.cs" Link="Utils\KnownResourceNames.cs" />
     <Compile Include="$(SharedDir)KnownFormats.cs" Link="Utils\KnownFormats.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <UpToDateCheckInput Remove="Components\ResourcesGridColumns\UnreadLogErrorsBadge.razor" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <_ContentIncludedByDefault Remove="Components\ResourcesGridColumns\UnreadLogErrorsBadge.razor" />
   </ItemGroup>
 
 </Project>
