+++ src/Aspire.Dashboard/Components/Pages/Resources.razor.css
@@ -7,10 +7,16 @@
     padding-left: calc(var(--design-unit) * 1px);
 }
 
+::deep .unread-logs-errors-link {
+    vertical-align: middle;
+    --unread-logs-badge-color: #ffffff;
+}
+
 ::deep .unread-logs-errors-link::part(control) {
-    padding-left: 0;
-    padding-right: 0;
-    margin-left: 3px;
+    margin-left: 5px;
+    padding: 0;
+    flex-grow: 0;
+    color: #ffffff;
 }
 
 ::deep .error-counter-badge {
