+++ src/Aspire.Dashboard/Components/ResourcesGridColumns/UnreadLogErrorsBadge.razor
@@ -1,0 +1,20 @@
+﻿@using Aspire.Dashboard.Resources
+@using Aspire.Dashboard.Model
+@namespace Aspire.Dashboard.Components
+@inject IStringLocalizer<Columns> Loc
+
+@if (_unviewedCount > 0)
+{
+    <FluentAnchor aria-label="@(_unviewedCount == 1 ? nameof(Columns.UnreadLogErrorsBadgeOneErrorLog) : string.Format(Loc[nameof(Columns.UnreadLogErrorsBadgeErrorLogs)], _unviewedCount))"
+                  Href="@GetResourceErrorStructuredLogsUrl(Resource)"
+                  Appearance="Appearance.Stealth"
+                  Class="unread-logs-errors-link">
+        <FluentBadge
+            Appearance="Appearance.Accent"
+            Circular="true"
+            Fill="error"
+            Color="var(--unread-logs-badge-color)">
+            @(_unviewedCount > 9 ? "9+" : _unviewedCount.ToString())
+        </FluentBadge>
+    </FluentAnchor>
+}
