+++ src/Aspire.Dashboard/Resources/Columns.resx
@@ -142,10 +142,6 @@
   <data name="SourceColumnSourceCopyContainerToClipboard" xml:space="preserve">
     <value>Copy image name and tag to clipboard</value>
   </data>
-  <data name="UnreadLogErrors" xml:space="preserve">
-    <value>{0} errors</value>
-    <comment>{0} is a number</comment>
-  </data>
   <data name="StateColumnResourceExitedUnexpectedly" xml:space="preserve">
     <value>{0} exited unexpectedly with exit code {1}</value>
     <comment>{0} is a resource type, {1} is a number</comment>
@@ -156,5 +152,12 @@
   </data>
   <data name="SourceColumnDisplayCopyCommandToClipboard" xml:space="preserve">
     <value>Copy full command to clipboard</value>
+  </data>
+  <data name="UnreadLogErrorsBadgeOneErrorLog" xml:space="preserve">
+    <value>1 error log</value>
+  </data>
+  <data name="UnreadLogErrorsBadgeErrorLogs" xml:space="preserve">
+    <value>{0} error logs</value>
+    <comment>{0} is a number</comment>
   </data>
 </root>
